Restructure data objects into models folder.
