+++ cli/cli-menu-structures/show-all-previous-active-fasts.js
@@ -1,4 +1,4 @@
-const MenuElement = require('../cli-menu-element');
+const MenuElement = require('../../data/models/cli-menu-element');
 const DataServiceSingleton = require('../../data/data-service');
 const {formatDatetimeString} = require('../datetime-helper');
