+++ data/models/fast.js
